+++ Fluxer.Net/Admin/Data/ManagedApplication.cs
@@ -1,0 +1,18 @@
+﻿namespace Fluxer.Net.Admin;
+
+public class ManagedApplication : Application
+{
+
+
+    internal ManagedApplication(FluxerBaseClient client) : base(client)
+    {
+
+    }
+
+    public static ManagedApplication Create(FluxerAdminClient client, ApplicationJson json)
+    {
+        ManagedApplication data = new ManagedApplication(client);
+        data.Update(json);
+        return data;
+    }
+}
