+++ Fluxer.Net/Fluxer.Net.csproj
@@ -51,5 +51,10 @@
 	<ItemGroup Condition="$(TargetFramework) == 'netstandard2.1'">
 		<PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
 	</ItemGroup>
+
+	<ItemGroup>
+	  <Folder Include="Admin\Helpers\" />
+	  <Folder Include="Admin\Requests\" />
+	</ItemGroup>
 </Project>
